+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/BaseBrowserFragment.kt
@@ -144,7 +144,6 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler {
         sitePermissionsFeature.set(
             feature = SitePermissionsFeature(
                 context = requireContext(),
-                sessionManager = components.sessionManager,
                 sessionId = sessionId,
                 storage = components.permissionStorage,
                 fragmentManager = requireFragmentManager(),